added population reading by population type
The functions getListofHistoricPop and getHistoricPop read the game's statistic history data for population per population type and print them to the debug.log file.

Caveat1: The last value from the statistic history is not necessarily the current population value as it can be about 5 minutes old. This is because the game takes snapshots every x minutes and adds then to the statistic history. But for many use cases that should be "close enough"

Caveat2: The functions read the population value from the current selected island (default: main island). To change the island you have to open the statistics menu and change the island on the left. Maybe there is a way to to than programatically?

getHistoricPop -> returns population value for a given population type (e.g. farmer) and a given snapshot (e.g. 2 = second oldest snapshot). You can check the values ingame in the statistics panel (Ctrl-Q -> population -> graph on the right)

getListofHistoricPop -> logs all population values from all snapshots for all population types specified in popIDs